@@ -75,48 +71,49 @@
     ]
   },
   "devDependencies": {
-    "@babel/core": "^7.12.13",
-    "@babel/eslint-parser": "^7.12.13",
+    "@babel/core": "^7.13.1",
+    "@babel/eslint-parser": "^7.13.4",
     "@babel/plugin-proposal-class-properties": "^7.12.1",
-    "@babel/preset-env": "^7.12.13",
+    "@babel/preset-env": "^7.13.5",
     "@babel/preset-react": "^7.12.13",
     "@mdx-js/loader": "^1.6.21",
     "@octokit/auth-action": "^1.3.2",
-    "@octokit/rest": "^18.1.0",
+    "@octokit/rest": "^18.2.1",
     "@pmmmwh/react-refresh-webpack-plugin": "^0.4.2",
     "@svgr/webpack": "^5.5.0",
     "alex": "^5.1.0",
     "autoprefixer": "^10.2.4",
     "babel-loader": "^8.2.2",
     "copy-webpack-plugin": "^7.0.0",
-    "css-loader": "^5.0.1",
+    "css-loader": "^5.0.2",
     "css-minimizer-webpack-plugin": "^1.2.0",
-    "cypress": "^6.4.0",
-    "directory-tree": "^2.2.6",
+    "cypress": "^6.5.0",
+    "directory-tree": "^2.2.7",
     "directory-tree-webpack-plugin": "^1.0.2",
     "duplexer": "^0.1.1",
-    "eslint": "^7.19.0",
-    "eslint-config-prettier": "^7.2.0",
+    "eslint": "^7.20.0",
+    "eslint-config-prettier": "^8.1.0",
     "eslint-plugin-cypress": "^2.11.2",
-    "eslint-plugin-markdown": "^2.0.0-rc.2",
+    "eslint-plugin-markdown": "^2.0.0",
     "eslint-plugin-react": "^7.22.0",
+    "eslint-plugin-react-hooks": "^4.2.0",
     "front-matter": "^4.0.2",
-    "html-loader": "^1.3.0",
-    "html-webpack-plugin": "^5.0.0",
+    "html-loader": "^2.1.1",
+    "html-webpack-plugin": "^5.2.0",
     "http-server": "^0.12.3",
-    "husky": "^4.3.8",
+    "husky": "^5.1.1",
     "hyperlink": "^4.6.0",
     "jest": "^26.6.3",
-    "lint-staged": "^10.5.3",
-    "lodash": "^4.17.19",
-    "markdownlint": "^0.22.0",
+    "lint-staged": "^10.5.4",
+    "lodash": "^4.17.21",
+    "markdownlint": "^0.23.1",
     "markdownlint-cli": "^0.26.0",
-    "mini-css-extract-plugin": "^1.3.5",
+    "mini-css-extract-plugin": "^1.3.8",
     "mkdirp": "^1.0.4",
     "modularscale-sass": "^3.0.3",
     "node-fetch": "^2.6.1",
     "npm-run-all": "^4.1.1",
-    "postcss": "^8.2.4",
+    "postcss": "^8.2.6",
     "postcss-loader": "^5.0.0",
     "prettier": "^2.2.1",
     "react-refresh": "^0.9.0",
@@ -141,12 +138,12 @@
     "style-loader": "^2.0.0",
     "tap-spot": "^1.1.1",
     "unist-util-visit": "^2.0.3",
-    "webpack": "^5.20.2",
+    "webpack": "^5.24.2",
     "webpack-bundle-analyzer": "^4.4.0",
     "webpack-cli": "^4.5.0",
     "webpack-dev-server": "^4.0.0-beta.0",
     "webpack-merge": "^5.7.3",
-    "workbox-webpack-plugin": "^6.1.0"
+    "workbox-webpack-plugin": "^6.1.1"
   },
   "dependencies": {
     "docsearch.js": "^2.5.2",
