@@ -14,12 +14,7 @@ const common = require('./webpack.common.js');
 const PrecacheSsgManifestPlugin = require('./src/PrecacheSsgManifestPlugin');
 
 // content tree to path array
-const paths = [
-  ...flattenContentTree(contentTree),
-  '/vote',
-  '/organization',
-  '/app-shell',
-];
+const paths = [...flattenContentTree(contentTree), '/vote', '/app-shell'];
 
 module.exports = (env) =>
   merge(common(env), {
@@ -29,6 +24,13 @@ module.exports = (env) =>
     cache: {
       buildDependencies: {
         config: [__filename],
+      },
+    },
+    module: {
+      parser: {
+        javascript: {
+          url: 'relative',
+        },
       },
     },
     entry: {
